Update tests to run with refactor; fix newton
