+++ dask_glm/tests/test_logistic.py
@@ -4,7 +4,7 @@ from dask import persist
 import numpy as np
 import dask.array as da
 
-from dask_glm.logistic import (newton, bfgs, proximal_grad,
+from dask_glm.algorithms import (newton, bfgs, proximal_grad,
                                gradient_descent, admm)
 from dask_glm.utils import sigmoid, make_y
